+++ test/v1_4_0/integration/merge/Points.t.sol
@@ -172,8 +172,8 @@ contract TestMerge_Points is IEscrowCurveTokenStorage, IEscrowCurveGlobalStorage
     }
 
     function testFuzz_Merge(
-        uint192 _lock1Amount,
-        uint192 _lock2Amount,
+        uint184 _lock1Amount,
+        uint184 _lock2Amount,
         uint48 _fromLockTime,
         uint48 _toLockTime,
         uint48 _mergeTime
