+++ test/v1_4_0/integration/split/Supply.t.sol
@@ -112,10 +112,10 @@ contract TestSplit_Supply is IEscrowCurveTokenStorage, IEscrowCurveGlobalStorage
     }
 
     function testFuzz_Split(
-        uint192 _lock1Amount,
-        uint192 _splitValue,
+        uint184 _lock1Amount,
+        uint184 _splitValue,
         uint48 _fromLockTime,
-        uint192 _splitTime
+        uint184 _splitTime
     ) public {
        (_fromLockTime, _splitTime) = boundLockCreationFuzzTimes(_fromLockTime, _splitTime);
